+++ Core/ModuleInstaller.cs
@@ -218,7 +218,7 @@ namespace CKAN
 
                 registry_manager.Save(!options.without_enforce_consistency);
 
-                User.RaiseProgress("Commiting filesystem changes", 80);
+                User.RaiseProgress("Committing filesystem changes", 80);
 
                 transaction.Complete();
 
@@ -255,7 +255,7 @@ namespace CKAN
 
                 registry_manager.Save(!options.without_enforce_consistency);
 
-                User.RaiseProgress("Commiting filesystem changes", 80);
+                User.RaiseProgress("Committing filesystem changes", 80);
 
                 transaction.Complete();
             }